requirements.txt
requests
python-dotenv
dash
plotly
pandas